# feat: convert `Node.Print` to use `switch (node.Kind())`