--- a/Src/CSharpier.Generators/NodePrinterGenerator.sbntxt
+++ b/Src/CSharpier.Generators/NodePrinterGenerator.sbntxt
@@ -1,4 +1,5 @@
 using System;
+using System.Diagnostics;
 using CSharpier.DocTypes;
 using CSharpier.SyntaxPrinter;
 using CSharpier.SyntaxPrinter.SyntaxNodePrinters;
@@ -29,13 +30,14 @@ namespace CSharpier.SyntaxPrinter
             context.State.PrintingDepth++;
             try
             {
-                switch (syntaxNode)
+                switch ((SyntaxKind)syntaxNode.RawKind)
                 {
                     {{- for nodeType in NodeTypes }}
-                    case {{ nodeType.SyntaxNodeName }} {{ nodeType.VariableName }}:
-                        return {{ nodeType.PrinterName }}.Print({{ nodeType.VariableName }}, context);
+                    case {{ nodeType.SyntaxKinds }}:
+                        return {{ nodeType.PrinterName }}.Print(({{ nodeType.SyntaxNodeName }})syntaxNode, context);
                     {{- end }}
                     default:
+                        Debug.Assert(syntaxNode is not ({{- SyntaxNodeTypes -}}));
                         return UnhandledNode.Print(syntaxNode, context);
                 }
             }