--- a/Src/CSharpier.Tests/FormattingTests/TestFiles/cs/LabeledStatements.test
+++ b/Src/CSharpier.Tests/FormattingTests/TestFiles/cs/LabeledStatements.test
@@ -3,6 +3,7 @@ class ClassName
     void MethodName()
     {
         label:
+
         goto label;
     }
 }