- Use `switch (syntaxNode.Kind())` instead of `is`. Should be slightly faster 🤞 
- Add `Debug.Assert` in the default branch to ensure that future changes to roslyn / mistakes with `SyntaxKind` are caught


### Benchmarks
Ran `FormatAsync` in a for loop 10 times to try and average performance. Seems to show a small 1.5% increase in performance at the cost of more complexity and the potential for errors. Do you think its worth merging this pr @belav ? TBH I'm a little dissapointed 😅 

I did wonder if using chained `or` might affect the generated code 🤷 the massive `SyntaxKind` switch statements in roslyn use several `case` tags instead. Perhaps this is lowered into better code? Looking at the IL I think it creates a jump table, although I'm not sure - I'm not fluent in IL 😄 .


#### Before
| Method                | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2      | Allocated |
|---------------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| Default_CodeFormatter | 2.115 s | 0.0132 s | 0.0103 s | 77000.0000 | 35000.0000 | 9000.0000 | 661.69 MB |

#### After
| Method                | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2      | Allocated |
|---------------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| Default_CodeFormatter | 2.085 s | 0.0168 s | 0.0149 s | 77000.0000 | 35000.0000 | 9000.0000 |  661.7 MB |