@TimothyMakkison 2025-02-26
> The special cases seem to be a bit painful to manage but I don't really see special cases being changed though, possibly just new ones added as the language evolves. The last few language versions have been really straightforward for CSharpier to support though and a few of them didn't even require any changes beyond updating the `CodeAnalysis` version. With the extra complexity hidden away I'm not super concerned with it.

Fair enough, for whats its worth I have manually checked each printer type case I missed any `SyntaxNode` types that don't share their name with a `SyntaxKind`. It's how I discovered that `unchecked` didn't have unit tests.

> I thought CSharpier's code coverage was solid until you uncovered the missing tests for `unchecked`. I do have a process for formatting everything in `csharpier-repos` which has caught quite a few bugs. I'm confident that would catch any errors this may introduce.

Did you change anything with CLI tests? 4 days ago they started failing. It keeps happening after I updated rider to 2024
![image](https://github.com/user-attachments/assets/c08f0d01-84e7-4e9b-85a0-846af6154299)

> There may be a tiny benefit to using [RawSyntaxKind](https://github.com/belav/csharpier/blob/main/Src/CSharpier/SyntaxTriviaExtensions.cs#L19) which I added a while back.

Good shout, it can't hurt to add it.

> I just heard of [FrozenDictionary](https://learn.microsoft.com/en-us/dotnet/api/system.collections.frozen.frozendictionary-2?view=net-9.0) today although with the dictionary you added only being in the generator I don't know how beneficial it would be.

Yeah, I've been looking for an excuse to use it for ages. Unfortunately it would be a bad fit for the generator 😞 . Briefly considered it for `Node.Print` but I'm pretty sure the compiler is faster.