@belav 2025-02-28
> Did you change anything with CLI tests? 4 days ago they started failing.

Nothing I can think of on `main`, most of my work lately has been going into `1.0.0`. I fixed a bug on `main` just now and that PR ran the tests and after updating this PR it is also running the tests.

Something I ran into on `1.0.0` is the CLI tests run inside of the bin of the project, which was pulling in a config file the tests weren't expecting. Maybe you have a file in your directory tree causing the same problem. I'm going to pull my `1.0.0` changes into `main` that move where the cli tests are run.

> Yeah, I've been looking for an excuse to use it for ages. Unfortunately it would be a bad fit for the generator 😞 . Briefly considered it for Node.Print but the compiler is faster.

Ah yeah good point. We do make use of it in a few spots at work, and maybe those would be better off as switch statements.