@belav 2025-02-28
It looks like there is something missing in this PR, all of these test files have a blank line before `goto` removed - https://github.com/belav/csharpier-repos/pull/127/files

I validated that the original of the first file before any formatting did have those blank lines - https://github.com/belav/csharpier-repos/blob/main/AspNetWebStack/src/System.Net.Http.Formatting/Formatting/Parsers/FormUrlEncodedParser.cs#L174