@TimothyMakkison 2025-02-28
mb I see the problem, `GotoStatementSyntax` has multiple kinds associated with it. I'll fix it and double check everything in case there are other issues.